1.异步对象模型增加LuaThread，就是协程 2.finalize 因为意义不明确。改成了terminate 3.对象池子的代码基本OK。 4.修改部分头文件的卫哨为#program once

+++ src/commlib/zcelib/zce/timer/queue_base.h
@@ -274,7 +274,7 @@ protected:
     ///普通服务器我建议你用s的级别，也就是1000ms
     ///要求精度比较高的服务器，我建议你用100ms
     ///要求不高的服务器，建议你用若干N s，N小于10s
-    unsigned int                timer_precision_mesc_ = DEFAULT_TIMER_PRECISION_MSEC;
+    uint64_t                    timer_precision_mesc_ = DEFAULT_TIMER_PRECISION_MSEC;
 
     ///已经使用的节点的数量，也就是分配了多少个定时器
     size_t                      num_use_node_ = 0;
